+++ src-tauri/build.rs
@@ -11,7 +11,7 @@ fn main() {
     if env::var("AUR_GENERATE_DEFS").is_ok()
         && env::var("AUR_PACKAGE_NAME").is_ok()
         && env::var("AUR_DEFS_OUTPUT").is_ok()
-        && env::var("AUR_DEB_PREFIX").is_ok()
+        && env::var("APP_BINARY_NAME").is_ok()
         && env::var("AUR_GITHUB_REPO").is_ok()
     {
         let crate_dir = PathBuf::from(env::var("CARGO_MANIFEST_DIR").unwrap());
@@ -19,7 +19,7 @@ fn main() {
         let aur_package_name = env::var("AUR_PACKAGE_NAME").unwrap();
 
         // Create dir
-        create_dir_all(&target_dir.join("build")).unwrap();
+        create_dir_all(&target_dir).unwrap();
 
         // Load AUR info and replace variables
         replace_file_vars(
@@ -34,18 +34,65 @@ fn main() {
             &target_dir.join(format!("{}.install", aur_package_name)),
         );
     }
+
+    // Generate Snapcraft.yml
+    if env::var("SNAP_GENERATE_DEFS").is_ok()
+        && env::var("SNAP_PACKAGE_NAME").is_ok()
+        && env::var("SNAP_DEFS_OUTPUT").is_ok()
+        && env::var("APP_BINARY_NAME").is_ok()
+        && env::var("SNAP_DEB_PATH").is_ok()
+    {
+        let crate_dir = PathBuf::from(env::var("CARGO_MANIFEST_DIR").unwrap());
+        let target_dir = PathBuf::from(env::var("SNAP_DEFS_OUTPUT").unwrap());
+
+        // Create dir
+        create_dir_all(&target_dir).unwrap();
+
+        // Load file and replace variables
+        replace_file_vars(
+            &crate_dir.join("packageinfo").join("snapcraft").join("snapcraft.yaml"),
+            &target_dir.join("snapcraft.yaml")
+        );
+
+    }
 }
 
 fn replace_file_vars(in_file: &PathBuf, out_file: &PathBuf) {
     let mut contents = fs::read_to_string(in_file).unwrap();
 
     let rules = &[
-        ("${AUR_PACKAGE_NAME}", env::var("AUR_PACKAGE_NAME").unwrap()),
-        ("${AUR_DEB_PREFIX}", env::var("AUR_DEB_PREFIX").unwrap()),
-        ("${AUR_GITHUB_REPO}", env::var("AUR_GITHUB_REPO").unwrap()),
-        ("${CARGO_PKG_VERSION}", env::var("CARGO_PKG_VERSION").unwrap()),
-        ("${CARGO_PKG_DESCRIPTION}", env::var("CARGO_PKG_DESCRIPTION").unwrap()),
-        ("${CARGO_PKG_HOMEPAGE}", env::var("CARGO_PKG_HOMEPAGE").unwrap()),
+        (
+            "${AUR_PACKAGE_NAME}",
+            env::var("AUR_PACKAGE_NAME").unwrap_or_default(),
+        ),
+        (
+            "${APP_BINARY_NAME}",
+            env::var("APP_BINARY_NAME").unwrap_or_default(),
+        ),
+        (
+            "${AUR_GITHUB_REPO}",
+            env::var("AUR_GITHUB_REPO").unwrap_or_default(),
+        ),
+        (
+            "${CARGO_PKG_VERSION}",
+            env::var("CARGO_PKG_VERSION").unwrap_or_default(),
+        ),
+        (
+            "${CARGO_PKG_DESCRIPTION}",
+            env::var("CARGO_PKG_DESCRIPTION").unwrap_or_default(),
+        ),
+        (
+            "${CARGO_PKG_HOMEPAGE}",
+            env::var("CARGO_PKG_HOMEPAGE").unwrap_or_default(),
+        ),
+        (
+            "${SNAP_PACKAGE_NAME}",
+            env::var("SNAP_PACKAGE_NAME").unwrap_or_default()
+        ),
+        (
+            "${SNAP_DEB_PATH}",
+            env::var("SNAP_DEB_PATH").unwrap_or_default()
+        )
     ];
 
     for (from, to) in rules {