+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.22.7] - 2025-02-11
+### Added
+- Snapcraft Build
+- AUR Build
+- Flathub Build
+
 ## [0.22.1] - 2025-02-08
 ### Added
 - ARM builds to windows and linux
@@ -85,7 +91,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Deleting Aircraft Bug
 
-[Unreleased]: https://github.com/sauna-sim/sauna-ui/compare/v0.22.1...master
+[Unreleased]: https://github.com/sauna-sim/sauna-ui/compare/v0.22.7...master
+[0.22.7]: https://github.com/sauna-sim/sauna-ui/compare/v0.22.1...v0.22.7
 [0.22.1]: https://github.com/sauna-sim/sauna-ui/compare/v0.21.11...v0.22.1
 [0.21.11]: https://github.com/sauna-sim/sauna-ui/compare/v0.21.10...v0.21.11
 [0.21.10]: https://github.com/sauna-sim/sauna-ui/compare/v0.21.9...v0.21.10